RasPi3_sequencer/requirements.txt
luma.core
luma.led_matrix
pygame
numpy
pyserial
spidev
sounddevice